[NativeSql][c++] Support date type
[NativeSql][Scala] support fall back row-based shuffle
